+++ ceammc/CHANGELOG.md
@@ -43,7 +43,7 @@
 - new object properties:
   - @punct added to rhvoice~, to specify pronounced characters
   - @bypass added to dyn.comp~ and dyn.comp2~
-  - @bypass added to dyn.gate~
+  - @bypass added to dyn.gate~ and dyn.gate2~
 - new init argument functions:
   - hexbeat(HEX) - expand hex to list of 1 and 0
   - ones(N) - list of length N, filled with 1